move series selector and its tests to their own files
